+++ lib/generators/sorcery/install_generator.rb
@@ -7,6 +7,16 @@ module Sorcery
       include Rails::Generators::Migration
       include Sorcery::Generators::Helpers
 
+      AVAILABLE_SUBMODULES = %w[
+        activity_logging
+        brute_force_protection
+        external
+        magic_login
+        remember_me
+        reset_password
+        user_activation
+      ].freeze
+
       source_root File.expand_path('templates', __dir__)
 
       argument :submodules, optional: true, type: :array, banner: 'submodules'
@@ -26,49 +36,52 @@ module Sorcery
         warn('[DEPRECATED] `--migrations` option is deprecated, please use `--only-submodules` instead')
       end
 
-      # Copy the initializer file to config/initializers folder.
-      def copy_initializer_file
-        template 'initializer.rb', sorcery_config_path unless only_submodules?
+      def check_available_submodules
+        return unless submodules
+
+        guidance = "Try some of these: #{AVAILABLE_SUBMODULES.join(', ')}"
+
+        submodules.each do |submodule|
+          raise ArgumentError, "#{submodule} is not a Sorcery submodule. #{guidance}" unless
+            AVAILABLE_SUBMODULES.include?(submodule)
+        end
       end
 
-      def configure_initializer_file
-        # Add submodules to the initializer file.
+      # Copy the initializer file to config/initializers folder, and add the submodules if necessary.
+      def install_initializer
+        template 'initializer.rb', initializer_path unless only_submodules?
+
         return unless submodules
 
         submodule_names = submodules.collect { |submodule| ':' + submodule }
 
-        gsub_file sorcery_config_path, /submodules = \[.*\]/ do |str|
+        gsub_file initializer_path, /submodules = \[.*\]/ do |str|
           current_submodule_names = (str =~ /\[(.*)\]/ ? Regexp.last_match(1) : '').delete(' ').split(',')
           "submodules = [#{(current_submodule_names | submodule_names).join(', ')}]"
         end
       end
 
-      def configure_model
+      def install_model
         # Generate the model and add 'authenticates_with_sorcery!' unless you passed --only-submodules
         return if only_submodules?
 
         generate "model #{model_class_name} --skip-migration"
-        inject_sorcery_to_model
-      end
-
-      def inject_sorcery_to_model
-        indents = '  ' * (namespaced? ? 2 : 1)
-
-        inject_into_class(model_path, model_class_name, "#{indents}authenticates_with_sorcery!\n")
+        
+        inject_into_class(model_path, model_class_name, model_injection, after: model_injection_point)
       end
 
       # Copy the migrations files to db/migrate folder
-      def copy_migration_files
+      def install_migrations
         # Copy core migration file in all cases except when you pass --only-submodules.
         return unless defined?(ActiveRecord)
 
-        migration_template 'migration/core.rb', 'db/migrate/sorcery_core.rb', migration_class_name: migration_class_name unless only_submodules?
+        migration_template 'migration/core.rb', migration_path(:core), migration_class_name: migration_class_name unless only_submodules?
 
         return unless submodules
 
         submodules.each do |submodule|
           unless %w[http_basic_auth session_timeout core].include?(submodule)
-            migration_template "migration/#{submodule}.rb", "db/migrate/sorcery_#{submodule}.rb", migration_class_name: migration_class_name
+            migration_template "migration/#{submodule}.rb", migration_path(submodule), migration_class_name: migration_class_name
           end
         end
       end
@@ -80,20 +93,6 @@ module Sorcery
           Time.new.utc.strftime('%Y%m%d%H%M%S')
         else
           format('%.3d', (current_migration_number(dirname) + 1))
-        end
-      end
-
-      private
-
-      def only_submodules?
-        options[:migrations] || options[:only_submodules]
-      end
-
-      def migration_class_name
-        if Rails::VERSION::MAJOR >= 5
-          "ActiveRecord::Migration[#{Rails::VERSION::MAJOR}.#{Rails::VERSION::MINOR}]"
-        else
-          'ActiveRecord::Migration'
         end
       end
     end